lezione 15: modificato nel database il tipo di dato delle colonne TARIFFA e COSTO_SOSTA, da FLOAT  a DECIMAL; modificato chiave nell'array di aggiunta parcheggio al db, nel controller del parcheggio; modifciato in index.php la conversione del valore della tariffa da INTVAL a FLOATVAL; modificato attributo name dell'input del nome dle parcheggio nel template dell'omonima pagina;

+++ lezione15/es_casa/index.php
@@ -56,7 +56,7 @@
 
     // aggiungi parcheggio
     if (isset($_POST['azione']) && $_POST['azione'] == 'registra-parcheggio') {
-        $parcheggio = new mParcheggio($_POST['nome'], intval($_POST['tariffa']));
+        $parcheggio = new mParcheggio($_POST['nome'], floatval($_POST['tariffa']));
         $cParcheggio->aggiungi_parcheggio($parcheggio);
     }
